rename: random 패키지 util 밖으로 이동
